Renamed test.js
